Parameters can now used to setup rules dynamically
