+++ HABApp/rule_manager/__init__.py
@@ -1,3 +1,2 @@
 from .rule_manager import RuleManager
 from .rule_file import RuleFile
-from .rule_parameters import RuleParameters
