+++ HABApp/runtime/runtime.py
@@ -4,6 +4,7 @@ import HABApp.config
 import HABApp.core
 import HABApp.rule_manager
 import HABApp.util
+import HABApp.parameters.parameters
 
 from .folder_watcher import FolderWatcher
 from .shutdown_helper import ShutdownHelper
@@ -51,9 +52,12 @@ class Runtime:
         self.mqtt_connection = HABApp.mqtt.MqttConnection(self.config.mqtt, self.shutdown)
         self.mqtt_connection.connect()
 
+        # Parameter Files
+        HABApp.parameters.parameters.setup(self.config, self.folder_watcher)
+
         # Rule engine
         self.rule_manager = HABApp.rule_manager.RuleManager(self)
-        self.rule_params = HABApp.rule_manager.RuleParameters(self.config, self.folder_watcher)
+
 
     @HABApp.util.PrintException
     def get_async(self):
